test: DRY tests and fix snaps
