+++ packages/article/__tests__/shared-with-style.native.js
@@ -9,7 +9,7 @@ import {
 } from "@times-components/jest-serializer";
 import { iterator } from "@times-components/test-utils";
 import "./mocks.native";
-import { snapshotTests } from "./shared.base.native";
+import { snapshotTests } from "./shared.base";
 
 export default () => {
   addSerializers(
